v1111 patch

+++ src/Core/Drawer.cpp
@@ -710,8 +710,8 @@ namespace LinaVG
             for (int i = vtxStart; i < buf->m_vertexBuffer.m_size; i++)
             {
                 auto& vtx = buf->m_vertexBuffer[i];
-                vtx.pos.x += position.x;
-                vtx.pos.y += position.y;
+                vtx.pos.x += static_cast<float>(static_cast<int>(position.x));
+                vtx.pos.y += static_cast<float>(static_cast<int>(position.y));
             }
         }
 
@@ -762,8 +762,8 @@ namespace LinaVG
             for (int i = vtxStart; i < buf->m_vertexBuffer.m_size; i++)
             {
                 auto& vtx = buf->m_vertexBuffer[i];
-                vtx.pos.x += position.x;
-                vtx.pos.y += position.y;
+                vtx.pos.x += static_cast<float>(static_cast<int>(position.x));
+                vtx.pos.y += static_cast<float>(static_cast<int>(position.y));
             }
         }
 
@@ -3133,8 +3133,8 @@ namespace LinaVG
         int            characterCount      = 0;
         bool           first               = true;
 
-        pos.x = static_cast<int>(pos.x);
-        pos.y = static_cast<int>(pos.y);
+        pos.x = static_cast<float>(static_cast<int>(pos.x));
+        pos.y = static_cast<float>(static_cast<int>(pos.y));
 
         auto drawChar = [&](TextCharacter& ch) {
             const int startIndex = buf->m_vertexBuffer.m_size;
